REF_CODE_REFACTOR: prelu refactoring the reference code for prelu operator.
BUG=refactoring existing code.
