+++ tensorflow/lite/micro/kernels/BUILD
@@ -162,6 +162,7 @@ cc_library(
         "pooling.cc",
         "pooling_common.cc",
         "prelu.cc",
+        "prelu_common.cc",
         "quantize.cc",
         "quantize_common.cc",
         "reduce.cc",
@@ -198,6 +199,7 @@ cc_library(
         "logistic.h",
         "micro_ops.h",
         "pooling.h",
+        "prelu.h",
         "quantize.h",
         "softmax.h",
         "svdf.h",
