+++ tensorflow/lite/micro/kernels/micro_ops.h
@@ -56,6 +56,7 @@ TfLiteRegistration Register_LOGICAL_AND();
 TfLiteRegistration Register_LOGICAL_OR();
 TfLiteRegistration Register_LOGISTIC();
 TfLiteRegistration Register_MAX_POOL_2D();
+TfLiteRegistration Register_PRELU();
 TfLiteRegistration Register_QUANTIZE();
 TfLiteRegistration Register_RELU();
 TfLiteRegistration Register_RELU6();
@@ -99,7 +100,6 @@ TfLiteRegistration Register_NOT_EQUAL();
 TfLiteRegistration Register_PACK();
 TfLiteRegistration Register_PAD();
 TfLiteRegistration Register_PADV2();
-TfLiteRegistration Register_PRELU();
 TfLiteRegistration Register_REDUCE_MAX();
 TfLiteRegistration Register_RESHAPE();
 TfLiteRegistration Register_RESIZE_NEAREST_NEIGHBOR();
