+++ tensorflow/lite/micro/micro_mutable_op_resolver.h
@@ -391,7 +391,7 @@ class MicroMutableOpResolver : public MicroOpResolver {
 
   TfLiteStatus AddPrelu() {
     return AddBuiltin(BuiltinOperator_PRELU,
-                      tflite::ops::micro::Register_PRELU(), ParsePrelu);
+                      tflite::Register_PRELU(), ParsePrelu);
   }
 
   TfLiteStatus AddQuantize() {
